+++ stdlib/provider.go
@@ -113,6 +113,7 @@ func (*stdlibProvider) Functions(_ context.Context) []func() function.Function {
 		slicefunc.NewMinNumberFunction,
 		slicefunc.NewMinStringFunction,
 		slicefunc.NewProductFunction,
+		slicefunc.NewRepeatFunction,
 		slicefunc.NewReplaceFunction,
 		slicefunc.NewSortListFunction,
 		stringfunc.NewCutFunction,